% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NineBoxPO.R
\name{create_NineBoxPO}
\alias{create_NineBoxPO}
\title{create_NineBoxPO}
\usage{
create_NineBoxPO(.pathPO, .path9Box, .year)
}
\arguments{
\item{.pathPO}{data path where official PO is saved}

\item{.path9Box}{data path where raw 9BOx data is saved}

\item{.year}{Year of interest}
}
\value{
.data a estandarize po data frame with added 9Box columns for current palcement and last placement
}
\description{
create_NineBoxPO
}
