% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_PerformancePO.R
\name{create_PerformancePO}
\alias{create_PerformancePO}
\title{create_PerformancePO}
\usage{
create_PerformancePO(.pathPO, .patPerformance, .year)
}
\arguments{
\item{.pathPO}{data path where official PO is saved}

\item{.patPerformance}{data path to raw performance data}

\item{.year}{Year of interest}
}
\value{
.data a estandarize po data frame with added performance columns for each choosen year
}
\description{
create_PerformancePO
}
