% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RFPerformance.R
\name{create_RFPerformance}
\alias{create_RFPerformance}
\title{create_RFPerformance}
\usage{
create_RFPerformance(.date, .dataPOPerformance)
}
\arguments{
\item{.date}{date of report}

\item{.dataPOPerformance}{standarized performance po}
}
\value{
RF Performance data
}
\description{
create_RFPerformance
}
