% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RFTalento.R
\name{create_RFTalento}
\alias{create_RFTalento}
\title{create_RFTalento}
\usage{
create_RFTalento(.date, .dataPONineBox)
}
\arguments{
\item{.date}{Date of the reporte}

\item{.dataPO}{standrarized official PO with data from 9NineBOx}
}
\value{
.data standarized data frame for Talento Reporting Factory
}
\description{
create_RFTalento
}
