% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_Performance.R
\name{error_Performance}
\alias{error_Performance}
\title{error_Performance}
\usage{
error_Performance(.data, .dataPO)
}
\arguments{
\item{.data}{Performance data from read_Performance.R}

\item{.dataPO}{Employee Standardized data}
}
\value{
KOF employees with multiple grades
}
\description{
error_Performance
}
