% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_YearNineBoxPO.R
\name{munge_YearNineBoxPO}
\alias{munge_YearNineBoxPO}
\title{munge_YearNineBoxPO}
\usage{
munge_YearNineBoxPO(.data9BOX, .dataPO, .year)
}
\arguments{
\item{.data9BOX}{standarized 9box data frame}

\item{.dataPO}{standarized HC data frame}

\item{.year}{year of 9Box evaulation}
}
\value{
.data standarized 9Box data frame with users id that exsits on PO for .year evaulation and .year-1 evaluation
}
\description{
munge_YearNineBoxPO
}
