% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_YearPerformance.R
\name{munge_YearPerformance}
\alias{munge_YearPerformance}
\title{munge_YearPerformance}
\usage{
munge_YearPerformance(.data, .year)
}
\arguments{
\item{.data}{Performance data from read_Performance.R}

\item{.year}{Year of interest}
}
\value{
Format wide datframe where each column is a year for grades and multiples
}
\description{
munge_YearPerformance
}
