% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineBoxCoordstoText.R
\name{NineBoxCoordstoText}
\alias{NineBoxCoordstoText}
\title{NineBoxCoordstoText}
\usage{
NineBoxCoordstoText(.x, .y)
}
\arguments{
\item{.x}{Potential   0,1,2 grade}

\item{.y}{Performance 0,1,2 grade}
}
\value{
a labeled grade for nine box
}
\description{
NineBoxCoordstoText
}
