% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NineBoxDataQuality.R
\name{NineBoxDataQuality}
\alias{NineBoxDataQuality}
\title{NineBoxDataQuality}
\usage{
NineBoxDataQuality(.path9Box, .pathPO, .year)
}
\arguments{
\item{.path9Box}{file where raw 9box data is saved}

\item{.pathPO}{file where PO standarized is saved}

\item{.year}{9box year of interest}
}
\value{
.data data frame where an employee have multiple 9box for the same year
}
\description{
NineBoxDataQuality
}
