% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NineBox.R
\name{read_NineBox}
\alias{read_NineBox}
\title{read_NineBox}
\usage{
read_NineBox(.path)
}
\arguments{
\item{.path}{file path where extracted 9Box reported is saved}
}
\value{
.data a standarized data frame
}
\description{
read_NineBox
}
