% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Performance.R
\name{read_Performance}
\alias{read_Performance}
\title{read_Performance}
\usage{
read_Performance(.path)
}
\arguments{
\item{.path}{path to Performance Employee Central Report}
}
\value{
data as dataframe with standarize tipes and field names.
}
\description{
read_Performance
}
